# Omits testing tools from Carthage build step